requirements.txt
# requirements.txt
fastapi
uvicorn
requests
mysql-connector-python
